refactor(api): move domain-logic helpers from lib to services
Realign src/lib and src/services per AGENTS.md placement rules
(domain logic lives in services/, generic helpers in lib/, *Service.ts
files belong in services/).

- welcomePageService.ts -> services/ (the only *Service.ts under lib/)
- articleExtractor.ts -> services/ (URL→HTML→Readability→Tiptap pipeline)
- youtubeExtractor.ts -> services/ (YouTube→Tiptap+AI summary pipeline)
- Move mirrored tests to src/__tests__/services/ and follow all imports

Generic helpers (clipServerFetch, clipUrlPolicy, cors, env, sentry,
videoServerExtension, youtubeEmbedServerExtension, etc.) remain in lib/.

tsc --noEmit and full vitest suite (1425 tests) pass.

Refs #1013
