+++ server/api/src/__tests__/services/welcomePageService.test.ts
