+++ server/api/src/agents/core/tools/fetchArticle.ts
@@ -16,7 +16,7 @@
  */
 import { tool } from "@langchain/core/tools";
 import { z } from "zod";
-import { extractArticleFromUrl, ClipFetchBlockedError } from "../../../lib/articleExtractor.js";
+import { extractArticleFromUrl, ClipFetchBlockedError } from "../../../services/articleExtractor.js";
 import { isClipUrlAllowedAfterDns } from "../../../lib/clipUrlPolicy.js";
 
 /** Tool name. */
