+++ server/api/src/lib/clipAndCreate.ts
@@ -13,7 +13,7 @@ import { prosemirrorJSONToYDoc } from "@tiptap/y-tiptap";
 import type { NodePgDatabase } from "drizzle-orm/node-postgres";
 import { pages, pageContents } from "../schema/index.js";
 import type * as schema from "../schema/index.js";
-import { buildArticleSchema, extractArticleFromUrl } from "./articleExtractor.js";
+import { buildArticleSchema, extractArticleFromUrl } from "../services/articleExtractor.js";
 import type { AIProviderType, TokenUsage } from "../types/index.js";
 import { ensureDefaultNote } from "../services/defaultNoteService.js";
