+++ server/api/src/routes/ingest.ts
@@ -25,7 +25,7 @@ import { rateLimit } from "../middleware/rateLimit.js";
 import { sources } from "../schema/sources.js";
 import { pageSources } from "../schema/pageSources.js";
 import { eq, and } from "drizzle-orm";
-import { extractArticleFromUrl } from "../lib/articleExtractor.js";
+import { extractArticleFromUrl } from "../services/articleExtractor.js";
 import { validateModelAccessOrThrow } from "../lib/aiAccessHelpers.js";
 import { callProvider, getProviderApiKeyName } from "../services/aiProviders.js";
 import { getUserTier } from "../services/subscriptionService.js";
